backend/pyproject.toml
[tool.poetry]
name = "tradesage"
version = "1.0.0"
description = "Flask backend for TradeSage - Advanced Crypto Trading Platform with backtesting and strategy implementation"
authors = ["imKXNNY <kenny.gueclue@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
start = "app:main"


[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
pandas = "^2.2.3"
numpy = "^2.2.3"
plotly = "^6.0.0"
ccxt = "^4.4.59"
pandas-ta = "^0.3.14b0"
python-dotenv = "^1.0.1"
backtrader = "^1.9.78.123"
markdown2 = "^2.5.3"
gunicorn = "^21.2.0"
flask-cors = "^5.0.0"
websocket-client = "^1.8.0"
google = "^3.0.0"
protobuf = "^5.29.3"
pytz = "^2025.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
black = "^24.1.1"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
